// src/common/enums/notification-type.enum.ts
export enum NotificationType {
  BOOKING_CREATED = 'BOOKING_CREATED',
  BOOKING_CONFIRMED = 'BOOKING_CONFIRMED',
  BOOKING_CANCELLED = 'BOOKING_CANCELLED',
  BOOKING_COMPLETED = 'BOOKING_COMPLETED',
  PAYMENT_SUCCESS = 'PAYMENT_SUCCESS',
  PAYMENT_FAILED = 'PAYMENT_FAILED',
  REFUND_INITIATED = 'REFUND_INITIATED',
  REVIEW_RECEIVED = 'REVIEW_RECEIVED',
  VENDOR_VERIFIED = 'VENDOR_VERIFIED',
  GENERAL = 'GENERAL',
}